cleanup: Duplicating ESC & / to 3rd thumb key resp.
This the start of a push to remove the outer columns on my Kyria.

I believe I can make it work pretty _OK_ with this - I'll probably end
up using the 3rd thumb keys as modifiers when held and esc or /. When
held I'll probably let the outer 6th columns 'move in' to the 5th
column. This way I can sort-of use the pinky for åä and '< while not
moving away too far from qwerty setup.

I am hesitant to go 'full hog' and have another layout all together
therefore I'm doing this. I'm too old to learn a completely new layout I
think.